updated titles of steps for donation progress pages, added step, handleamountchnage, and amount in fuction
