+++ src/components/DonationPage/AddressForm.js
@@ -9,7 +9,7 @@ export default function AddressForm() {
   return (
     <React.Fragment>
       <Typography variant="h6" gutterBottom>
-        Shipping address
+       Donation Information
       </Typography>
       <Grid container spacing={3}>
         <Grid item xs={12} sm={6}>
